📝 feat. 소켓연결
